refactor(shared): narrow reasoning exports
