+++ sdk/packages/shared/src/index.browser.ts
@@ -194,13 +194,8 @@ export {
 	resolveReasoningEffortRatio,
 } from "./llms/reasoning-effort";
 export {
-	getModelReasoningControls,
-	type ModelReasoningControls,
-	type ModelReasoningEffortValue,
-	ModelReasoningEffortValueSchema,
 	type ModelReasoningOption,
 	ModelReasoningOptionSchema,
-	normalizeReasoningEffort,
 	REASONING_LEVELS,
 	type ReasoningEffort,
 	ReasoningEffortSchema,
